Move typed-constants inside contango
